[MOV] ranges: move adapt range to helper
