+++ src/plugins/core/range.ts
@@ -1,13 +1,9 @@
 import { compile } from "../../formulas";
 import {
-  createAdaptedZone,
   duplicateRangeInDuplicatedSheet,
+  getApplyRangeChange,
   getCanonicalSymbolName,
-  groupConsecutive,
-  isZoneInside,
   isZoneValid,
-  largeMax,
-  largeMin,
   numberToLetters,
   RangeImpl,
   rangeReference,
@@ -20,7 +16,6 @@ import { CellErrorType } from "../../types/errors";
 import {
   ApplyRangeChange,
   ApplyRangeChangeResult,
-  ChangeType,
   Command,
   CommandHandler,
   CommandResult,
@@ -76,143 +71,9 @@ export class RangeAdapter implements CommandHandler<CoreCommand> {
   beforeHandle(command: Command) {}
 
   handle(cmd: Command) {
-    switch (cmd.type) {
-      case "REMOVE_COLUMNS_ROWS": {
-        let start: "left" | "top" = cmd.dimension === "COL" ? "left" : "top";
-        let end: "right" | "bottom" = cmd.dimension === "COL" ? "right" : "bottom";
-        let dimension: "columns" | "rows" = cmd.dimension === "COL" ? "columns" : "rows";
-
-        const elements = [...cmd.elements];
-        elements.sort((a, b) => b - a);
-
-        const groups = groupConsecutive(elements);
-        this.executeOnAllRanges((range: RangeImpl) => {
-          if (range.sheetId !== cmd.sheetId) {
-            return { changeType: "NONE" };
-          }
-          let newRange = range;
-          let changeType: ChangeType = "NONE";
-          for (let group of groups) {
-            const min = largeMin(group);
-            const max = largeMax(group);
-            if (range.zone[start] <= min && min <= range.zone[end]) {
-              const toRemove = Math.min(range.zone[end], max) - min + 1;
-              changeType = "RESIZE";
-              newRange = this.createAdaptedRange(newRange, dimension, changeType, -toRemove);
-            } else if (range.zone[start] >= min && range.zone[end] <= max) {
-              changeType = "REMOVE";
-              newRange = range.clone({ ...this.getInvalidRange() });
-            } else if (range.zone[start] <= max && range.zone[end] >= max) {
-              const toRemove = max - range.zone[start] + 1;
-              changeType = "RESIZE";
-              newRange = this.createAdaptedRange(newRange, dimension, changeType, -toRemove);
-              newRange = this.createAdaptedRange(
-                newRange,
-                dimension,
-                "MOVE",
-                -(range.zone[start] - min)
-              );
-            } else if (min < range.zone[start]) {
-              changeType = "MOVE";
-              newRange = this.createAdaptedRange(newRange, dimension, changeType, -(max - min + 1));
-            }
-          }
-          if (changeType !== "NONE") {
-            return { changeType, range: newRange };
-          }
-          return { changeType: "NONE" };
-        }, cmd.sheetId);
-        break;
-      }
-      case "ADD_COLUMNS_ROWS": {
-        let start: "left" | "top" = cmd.dimension === "COL" ? "left" : "top";
-        let end: "right" | "bottom" = cmd.dimension === "COL" ? "right" : "bottom";
-        let dimension: "columns" | "rows" = cmd.dimension === "COL" ? "columns" : "rows";
-
-        this.executeOnAllRanges((range: RangeImpl) => {
-          if (range.sheetId !== cmd.sheetId) {
-            return { changeType: "NONE" };
-          }
-          if (cmd.position === "after") {
-            if (range.zone[start] <= cmd.base && cmd.base < range.zone[end]) {
-              return {
-                changeType: "RESIZE",
-                range: this.createAdaptedRange(range, dimension, "RESIZE", cmd.quantity),
-              };
-            }
-            if (cmd.base < range.zone[start]) {
-              return {
-                changeType: "MOVE",
-                range: this.createAdaptedRange(range, dimension, "MOVE", cmd.quantity),
-              };
-            }
-          } else {
-            if (range.zone[start] < cmd.base && cmd.base <= range.zone[end]) {
-              return {
-                changeType: "RESIZE",
-                range: this.createAdaptedRange(range, dimension, "RESIZE", cmd.quantity),
-              };
-            }
-            if (cmd.base <= range.zone[start]) {
-              return {
-                changeType: "MOVE",
-                range: this.createAdaptedRange(range, dimension, "MOVE", cmd.quantity),
-              };
-            }
-          }
-          return { changeType: "NONE" };
-        }, cmd.sheetId);
-
-        break;
-      }
-      case "DELETE_SHEET": {
-        this.executeOnAllRanges((range: RangeImpl) => {
-          if (range.sheetId !== cmd.sheetId) {
-            return { changeType: "NONE" };
-          }
-          const invalidSheetName = this.getters.getSheetName(cmd.sheetId);
-          range = range.clone({
-            ...this.getInvalidRange(),
-            invalidSheetName,
-          });
-          return { changeType: "REMOVE", range };
-        }, cmd.sheetId);
-
-        break;
-      }
-      case "RENAME_SHEET": {
-        this.executeOnAllRanges((range: RangeImpl) => {
-          if (range.sheetId === cmd.sheetId) {
-            return { changeType: "CHANGE", range };
-          }
-          if (cmd.name && range.invalidSheetName === cmd.name) {
-            const invalidSheetName = undefined;
-            const sheetId = cmd.sheetId;
-            const newRange = range.clone({ sheetId, invalidSheetName });
-            return { changeType: "CHANGE", range: newRange };
-          }
-          return { changeType: "NONE" };
-        });
-        break;
-      }
-      case "MOVE_RANGES": {
-        const originZone = cmd.target[0];
-        this.executeOnAllRanges((range: RangeImpl) => {
-          if (range.sheetId !== cmd.sheetId || !isZoneInside(range.zone, originZone)) {
-            return { changeType: "NONE" };
-          }
-          const targetSheetId = cmd.targetSheetId;
-          const offsetX = cmd.col - originZone.left;
-          const offsetY = cmd.row - originZone.top;
-          const adaptedRange = this.createAdaptedRange(range, "both", "MOVE", [offsetX, offsetY]);
-          const prefixSheet = cmd.sheetId === targetSheetId ? adaptedRange.prefixSheet : true;
-          return {
-            changeType: "MOVE",
-            range: adaptedRange.clone({ sheetId: targetSheetId, prefixSheet }),
-          };
-        });
-        break;
-      }
+    const arc = getApplyRangeChange(cmd, this.getters);
+    if (arc?.applyChange) {
+      this.executeOnAllRanges(arc.applyChange, arc.sheetId);
     }
   }
 
@@ -232,17 +93,6 @@ export class RangeAdapter implements CommandHandler<CoreCommand> {
       }
       return result;
     };
-  }
-
-  private createAdaptedRange<Dimension extends "columns" | "rows" | "both">(
-    range: RangeImpl,
-    dimension: Dimension,
-    operation: "MOVE" | "RESIZE",
-    by: Dimension extends "both" ? [number, number] : number
-  ) {
-    const zone = createAdaptedZone(range.unboundedZone, dimension, operation, by);
-    const adaptedRange = range.clone({ zone });
-    return adaptedRange;
   }
 
   private executeOnAllRanges(adaptRange: ApplyRangeChange, sheetId?: UID) {
@@ -582,15 +432,5 @@ export class RangeAdapter implements CommandHandler<CoreCommand> {
     }
 
     return str;
-  }
-
-  private getInvalidRange() {
-    return {
-      parts: [],
-      prefixSheet: false,
-      zone: { left: -1, top: -1, right: -1, bottom: -1 },
-      sheetId: "",
-      invalidXc: CellErrorType.InvalidReference,
-    };
   }
 }
